feat/reservations_app_refactor
