+++ bookings/apps.py
@@ -3,7 +3,7 @@ from django.apps import AppConfig
 
 class ReservationConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'reservation'
+    name = 'reservations'
 
     def ready(self):
         from . import signals
